Added optional request parameter to as_html.
If the table happens to be rendered in Python (we are using the table to render a multiple choice widget), then the {% render_table %} template tag is difficult to use and we used as_html(). However, this clobbers the links. Being able to supply the actual request to the as_html() method solves this issue.