Make Meta.fields = () mean take no columns from the model fields. refs #183
Previously any non-truthy value meant take all columns from the model
fields. This prevented changing a column autogenerated from a model field
in a base class from being overridden in a derived class with a custom
column. If the column was left in Meta.fields the autogenerated column
would overwrite the custom column, so you had to remove it from
Meta.fields. However if that was the last column in Meta.fields it would
then be empty, including all model columns.

by @mbertheau