BooleanColumn: Test case that breaks with choice=... parameter
Adds test_boolean_field_choices_with_real_model_instances() to
tests/columns/test_booleancolumn.py.

After a wild-goose chase suspecting Django localization (which was
apparently not the problem), this is a test case that will fail for a
BooleanColumn representing a model.BooleanField with a
choices=... parameter set.

The reason that this test fails seems to be somewhere in the code path
that handles real model data. The test will pass when just using a pair of
mock-up dictionaries.

The test will also pass when removing the choices=... parameter from the
field = models.BooleanField(...) declaration.