Update table_data.rst with "sequence" and "exclude" examples
Adding an example on this page of the tutorial will be helpful for beginners, who might struggle to realize how to use the "sequence", "include" and "exclude" options.

Even though the previous page of the tutorial explains a brief "include" example, this might not be obvious to a beginner. The tutorial also doesn't explain how each of these options works and this example helps to visualize it in a brief way.